core: commonwealth: Move InterceptHandler to common utils
