+++ core/libs/commonwealth/commonwealth/utils/logs.py
@@ -1,0 +1,25 @@
+import logging
+from logging import LogRecord
+from types import FrameType
+from typing import Optional, Union
+
+from loguru import logger
+
+
+class InterceptHandler(logging.Handler):
+    def emit(self, record: LogRecord) -> None:
+        # Get corresponding Loguru level if it exists
+        level: Union[int, str]
+        try:
+            level = logger.level(record.levelname).name
+        except ValueError:
+            level = record.levelno
+
+        # Find caller from where originated the logged message
+        frame: Optional[FrameType]
+        frame, depth = logging.currentframe(), 2
+        while frame and frame.f_code.co_filename == logging.__file__:
+            frame = frame.f_back
+            depth += 1
+
+        logger.opt(depth=depth, exception=record.exc_info).log(level, record.getMessage())
