+++ core/services/ardupilot_manager/setup.py
@@ -59,5 +59,6 @@ setuptools.setup(
         "fastapi-versioning == 0.9.0",
         "aiofiles == 0.6.0",
         "loguru == 0.5.3",
+        "commonwealth == 0.1.0",
     ],
 )
